Address review feedback: Remove hybrid approach, use direct migration
Per reviewer feedback, removed the .legacy.tsx and .new.tsx files and
migrated components directly in place. This simplifies the migration by:

- Replacing ShowKeyboardShortcuts.tsx with plain CSS implementation
- Replacing KeyboardShortcutsPopup.tsx with plain CSS implementation
- Removing intermediate .legacy.tsx and .new.tsx files
- Maintaining all functionality with CSS variables and plain CSS

All styled-components have been replaced with plain CSS classes and
CSS variables for dynamic theming.

Address review feedback: Remove hybrid approach, use direct migration

Per reviewer feedback, removed the .legacy.tsx and .new.tsx files and
migrated components directly in place. This simplifies the migration by:

- Replacing ShowKeyboardShortcuts.tsx with plain CSS implementation
- Replacing KeyboardShortcutsPopup.tsx with plain CSS implementation
- Removing intermediate .legacy.tsx and .new.tsx files
- Maintaining all functionality with CSS variables and plain CSS

All styled-components have been replaced with plain CSS classes and
CSS variables for dynamic theming.

snaps

Replace 'any' types with proper Omit utility types

Per review feedback, replaced all 'as any' type assertions with
proper TypeScript Omit utility types. This provides better type
safety while still filtering out the theme prop to prevent it
from being passed to DOM elements.

Changes:
- ShowKeyboardShortcuts.tsx: 7 instances of 'as any' replaced
- KeyboardShortcutsPopup.tsx: 1 instance of 'as any' replaced

Each component now uses 'as Omit<PropType, 'className' | ...>'
to properly type the destructured props while excluding the
props that were already extracted in the function parameters.
