--- src/app/content/keyboardShortcuts/components/KeyboardShortcutsPopup.legacy.tsx
@@ -1,7 +1,0 @@
-/**
- * KeyboardShortcutsPopup - Legacy styled-components export
- *
- * For backward compatibility, we simply re-export from the .new file.
- * The only styled component was StyledModal which is now handled with a plain CSS class.
- */
-export { default } from './KeyboardShortcutsPopup.new';
